refactor(wiz-ui-react): move button to buttons
