+++ packages/wiz-ui-react/src/components/base/index.ts
@@ -1,1 +1,1 @@
-export * from "./button";
+export * from "./buttons";
